refactor(checker): split impossible-union pruning out of lazy.rs to hold the 2000-LOC boundary
Wiring the cache-poisoning backstop pushed `lazy.rs` to 2017 lines, over the
checker 2000-LOC arch boundary. Move the cohesive "impossible object-union
member pruning with env" cluster (`prune_impossible_object_union_members_with_env`
and its helpers) into a sibling `lazy_impossible_pruning` module. Pure code
move — same `CheckerState` impl, no behavior change. `lazy.rs` is now 1875 lines.
