+++ crates/tsz-checker/src/state/type_environment/mod.rs
@@ -8,6 +8,7 @@ mod formatting;
 pub(crate) mod lazy;
 mod lazy_flow_mirror;
 mod lazy_fuel;
+mod lazy_impossible_pruning;
 mod property_access_visited;
 mod source_location;
 mod type_node_resolution;
